npm + typescript fix + vue-typer and png typings
