+++ client/src/types/shims-typer.d.ts
@@ -1,0 +1,1 @@
+declare module "vue-typer";
